+++ src/frontend/src/lib/styles/tailwind/colors/base.js
